pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[[tool.uv.index]]
name = "aliyun"
url = "https://mirrors.aliyun.com/pypi/simple/"
default = true

[tool.uv]
default-groups = ["dev"]

[project]
name = "mplang"
dynamic = ["version"]
description = "Multi-Party Programming Language"
readme = "README.md"
license = { file = "LICENSE" }
authors = [
    { name = "SecretFlow Team", email = "secretflow-contact@service.alipay.com" },
]
requires-python = ">=3.10,<3.12"
dependencies = [
    "protobuf>=5.0,<6.0",
    "typing_extensions",
    # both FE & BE
    "spu>=0.9.4",
    "pyarrow>=14.0.0",
    # runtime only
    "lightphe>=0.0.15,<0.1.0", # for PHE backend implementation
    "tenseal==0.3.16",         # for FHE(SEAL) backend implementation
    "pandas>=2.0.0",
    "duckdb>=1.0.0",
    "httpx>=0.27.0",
    "fastapi",
    "uvicorn[standard]",
    "sqlglot>=23.0.0",
    "cryptography>=43.0.0",
    "coincurve>=20.0.0",
    "PyYAML>=6.0",
]

[project.scripts]
mplang-cli = "mplang.runtime.cli:main"

[dependency-groups]
dev = ["pytest", "pytest-cov", "pytest-asyncio", "ruff", "mypy", "pre-commit"]

examples = [
    "scikit-learn",
    "tensorflow",
    "xgboost>=2.0.0,<3.0.0", # CPU version, 3.x requires CUDA dependencies
]

[tool.hatch.metadata.hooks.custom]

[tool.hatch.build.targets.wheel]
packages = ["mplang"]

[tool.hatch.build.targets.sdist]
include = [
    "/mplang",
    "/scripts",
    "/examples",
    "/tests",
    "/tutorials",
    "/README.md",
    "/LICENSE",
    "/pyproject.toml",
]

[tool.hatch.version]
source = "vcs"

[tool.hatch.version.raw-options]
local_scheme = "no-local-version"

[tool.hatch.envs.default]
installer = "uv"

[tool.ruff]
# Use recommended settings with black-compatible formatting
line-length = 88
target-version = "py310"
# Enable preview mode for latest formatting improvements
preview = true

# Exclude generated files and protobuf code
exclude = [
    "mplang/protos/",    # Protobuf generated Python files
    "protos/",           # Protobuf source files
    "**/protos/**/*.py", # Any protobuf files in protos directories
    "**/*_pb2.py",       # Generated protobuf message files
    "**/*_pb2_grpc.py",  # Generated protobuf gRPC files
    "**/*_pb2.pyi",      # Generated protobuf type stubs
    "**/*_pb2_grpc.pyi", # Generated protobuf gRPC type stubs
]

[tool.ruff.lint]
# Use recommended rule set
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "ICN", # flake8-import-conventions
    "PIE", # flake8-pie
    "PYI", # flake8-pyi
    "RSE", # flake8-raise
    "RUF", # ruff-specific rules
]

# Ignore some rules that cause too much noise initially
ignore = [
    "E501",   # line-too-long (will be handled by formatter when possible)
    "E731",   # lambda-assignment (can be acceptable in some cases)
    "RUF002", # ambiguous-unicode-character-docstring (may be intentional)
]

# Per-file ignores for specific cases
[tool.ruff.lint.per-file-ignores]
"examples/**/*.py" = [
    "B023",
] # Allow intentional loop variable binding in examples
"tutorials/**/*.py" = [
    "B023",
] # Allow intentional loop variable binding in tutorials
"tutorials/pitfalls/*.py" = [
    "B023",
] # These files intentionally demonstrate pitfalls
"tests/**/*.py" = ["B017"] # Allow generic exception catching in tests

[tool.ruff.format]
# Use black-compatible formatting
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true           # Format code in docstrings
docstring-code-line-length = "dynamic" # Use same line length for docstring code

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disable_error_code = ["import-untyped"]

[[tool.mypy.overrides]]
# ignore errors in generated protobuf files
module = ["mplang.protos.v1alpha1.*"]
ignore_errors = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--strict-markers --strict-config"
markers = ["integration: cross-module or multi-party integration tests"]
filterwarnings = [
    # Suppress insecure mock TEE warnings in tests
    "ignore:Insecure mock TEE kernel 'mock_tee.quote_gen' in use.*:UserWarning",
    "ignore:Insecure mock TEE kernel 'mock_tee.attest' in use.*:UserWarning",
]
